requirements.txt
# Dependências Core
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.21.0

# Exportação de Dados
openpyxl>=3.0.9
reportlab>=3.6.0
Pillow>=8.3.0

# Banco de Dados MySQL
mysql-connector-python>=8.0.0
